FIX: extend incode processor
ADD: day 09 solution

+++ src/main/java/de/schoradt/adventofcode2019/Day02.java
@@ -20,7 +20,7 @@ public class Day02 extends Day00 {
         
         processor.init(noun, verb);
         
-        return processor.process();
+        return Long.valueOf(processor.process()).intValue();
     }
     
     public int hackNounVerb(String line, int result) {
